move files into components and pages dirs.
