+++ pages/homepage.jsx
@@ -2,8 +2,7 @@ var React = require('react');
 var Router = require('react-router');
 var Link = Router.Link;
 
-var Page = require('./page.jsx');
-var HeroUnit = require('./hero-unit.jsx');
+var HeroUnit = require('../components/hero-unit.jsx');
 
 var CaseStudies = React.createClass({
   render: function() {
